corrected issues after first review
