+++ pom.xml
@@ -3,7 +3,7 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>by.mus</groupId>
+    <groupId>by.foo</groupId>
     <artifactId>renamer</artifactId>
     <version>-</version>
     <packaging>jar</packaging>
@@ -32,7 +32,7 @@
                     <finalName>${project.artifactId}</finalName>
                     <archive>
                         <manifest>
-                            <mainClass>Main</mainClass>
+                            <mainClass>by.foo.Main</mainClass>
                         </manifest>
                     </archive>
                     <descriptorRefs>
