+++ src/main/java/by/foo/Main.java
@@ -1,14 +1,18 @@
+package by.foo;
+
+import by.foo.inspector.SongInspector;
 import ealvatag.audio.exceptions.CannotReadException;
 import ealvatag.audio.exceptions.InvalidAudioFrameException;
 import ealvatag.tag.*;
 
 import java.io.IOException;
+import java.net.URISyntaxException;
 
 
 public class Main {
-    public static void main(String[] args) throws IOException, TagException, InvalidAudioFrameException, CannotReadException {
-        new CurrentDirectory();
-    }
+  public static void main(String[] args) {
+    new SongInspector();
+  }
 }
